+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Word formatter generating substrings of the word.
 - Word formatter applying other word formatter in series. The output for the first word formatter is used as input for the second word formatter, etc.
 - Word formatter filtering words by length.
+- Word formatter filtering unique words.
 
 ### Changed
 - Word converters have been renamed word formatters.